refactor(game): Rename Circle to Player and update related variables
The Circle class and its associated variables were too generic to represent the player's character in the game. To make the code more self-explanatory, the following changes were made:

1. Renamed `Circle` class to `Player`.
2. Renamed `circleRadius` to `playerRadius`.
3. Renamed `circleSpeed` to `playerSpeed`.

These changes will make it easier to understand the role of the Player class and its associated attributes in the game's context. It also lays the groundwork for introducing other types of entities that may interact with the player.